test: add test classes for dependency injection in TestBehavior
