--- test/Core/Mocks/TestBehaviorOtherType.ts
@@ -1,18 +1,0 @@
-import { Behavior } from "@core/Behavior.ts";
-
-export class TestBehaviorOtherType extends Behavior {
-  public enableCount = 0;
-  public disableCount = 0;
-  public tickCount = 0;
-  public onEnable() {
-    this.enableCount++;
-  }
-
-  public onDisable() {
-    this.disableCount++;
-  }
-
-  public tick(deltaTime: number) {
-    this.tickCount++;
-  }
-}
